reorg/refactor: pull out fixtures, common set-up
