+++ spec/fixtures/empty_aasm.rb
@@ -1,0 +1,5 @@
+require 'aasm'
+
+class EmptyAasm < ActiveRecord::Base
+  include AASM
+end
