+++ spec/fixtures/no_aasm.rb
@@ -1,0 +1,2 @@
+class NoAasm
+end
