+++ spec/fixtures/single_state.rb
@@ -1,0 +1,12 @@
+require 'aasm'
+
+class SingleState < ActiveRecord::Base
+  include AASM
+
+  aasm do
+    state :single,
+          initial: true,
+          enter: [:foo, :bar],
+          exit: [:baz, :quux]
+  end
+end
